Rename ESAliasesException to TooManyAliasesException
Also move to a standalone class.
