+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/DeflectorResource.java
@@ -24,7 +24,7 @@ import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.graylog2.auditlog.jersey.AuditLog;
 import org.graylog2.configuration.ElasticsearchConfiguration;
 import org.graylog2.indexer.Deflector;
-import org.graylog2.indexer.indices.Indices;
+import org.graylog2.indexer.indices.TooManyAliasesException;
 import org.graylog2.plugin.cluster.ClusterConfigService;
 import org.graylog2.plugin.indexer.rotation.RotationStrategy;
 import org.graylog2.rest.models.system.deflector.responses.DeflectorSummary;
@@ -76,7 +76,7 @@ public class DeflectorResource extends RestResource {
     @ApiOperation(value = "Get current deflector status")
     @RequiresPermissions(RestPermissions.DEFLECTOR_READ)
     @Produces(MediaType.APPLICATION_JSON)
-    public DeflectorSummary deflector() throws ClassNotFoundException, Indices.ESAliasesException {
+    public DeflectorSummary deflector() throws ClassNotFoundException, TooManyAliasesException {
         return DeflectorSummary.create(deflector.isUp(), deflector.getCurrentActualTargetIndex());
     }
