+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/indexer/IndicesResource.java
@@ -31,6 +31,7 @@ import org.graylog2.indexer.Deflector;
 import org.graylog2.indexer.cluster.Cluster;
 import org.graylog2.indexer.indices.IndexStatistics;
 import org.graylog2.indexer.indices.Indices;
+import org.graylog2.indexer.indices.TooManyAliasesException;
 import org.graylog2.rest.models.system.indexer.requests.IndicesReadRequest;
 import org.graylog2.rest.models.system.indexer.responses.AllIndices;
 import org.graylog2.rest.models.system.indexer.responses.ClosedIndices;
@@ -219,7 +220,7 @@ public class IndicesResource extends RestResource {
         @ApiResponse(code = 403, message = "You cannot close the current deflector target index.")
     })
     @AuditLog(action = "closed", object = "index")
-    public void close(@ApiParam(name = "index") @PathParam("index") @NotNull String index) throws Indices.ESAliasesException {
+    public void close(@ApiParam(name = "index") @PathParam("index") @NotNull String index) throws TooManyAliasesException {
         checkPermission(RestPermissions.INDICES_CHANGESTATE, index);
 
         if (!deflector.isGraylogIndex(index)) {
@@ -244,7 +245,7 @@ public class IndicesResource extends RestResource {
         @ApiResponse(code = 403, message = "You cannot delete the current deflector target index.")
     })
     @AuditLog(object = "index")
-    public void delete(@ApiParam(name = "index") @PathParam("index") @NotNull String index) throws Indices.ESAliasesException {
+    public void delete(@ApiParam(name = "index") @PathParam("index") @NotNull String index) throws TooManyAliasesException {
         checkPermission(RestPermissions.INDICES_DELETE, index);
 
         if (!deflector.isGraylogIndex(index)) {
